#165:HeartRatingを作成
